Moved parser functions into it's own file
This is to seperate out the parser functionality from everything else
